requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
appdirs==1.4.4
async-timeout==4.0.2
asynctest==0.13.0
attrs==22.1.0
audioread==3.0.0
beautifulsoup4==4.11.1
bibtexparser==1.2.0
bs4==0.0.1
certifi==2021.10.8
cffi==1.15.1
charset-normalizer==2.0.12
cycler==0.11.0
datasets==2.4.0
decorator==5.1.1
dill==0.3.5.1
efficientnet-pytorch==0.7.0
ffmpeg-python==0.2.0
filelock==3.8.0
fonttools==4.33.3
frozenlist==1.3.1
fsspec==2022.7.1
future==0.18.2
huggingface-hub==0.9.1
idna==3.3
importlib-metadata==4.12.0
joblib==1.1.0
kiwisolver==1.4.2
librosa==0.9.2
llvmlite==0.38.0
matplotlib==3.5.2
multidict==6.0.2
multiprocess==0.70.13
numba==0.55.1
numpy==1.21.6
opencv-python==4.6.0.66
packaging==21.3
pandas==1.3.5
Pillow==9.1.0
pooch==1.6.0
pyarrow==9.0.0
pycparser==2.21
pyparsing==3.0.8
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
requests==2.27.1
resampy==0.4.0
responses==0.18.0
scikit-learn==0.24.2
scipy==1.6.3
six==1.16.0
sklearn==0.0
SoundFile==0.10.3.post1
soupsieve==2.3.2.post1
threadpoolctl==3.1.0
timm==0.4.5
torch==1.11.0
torchaudio==0.11.0
torchvision==0.12.0
tqdm==4.64.0
typing_extensions==4.2.0
urllib3==1.26.9
wget==3.2
xxhash==3.0.0
yarl==1.8.1
zipp==3.4.1
